fixed routes, views and error cypress tests
