+++ cypress/e2e/dev/7-error-page-tests/server-error.cypress.js
@@ -1,31 +1,43 @@
+const path = require('path')
 const { waitForApplication } = require('../../utils')
-const { setRouterBackToInitialState, setupRouterForErrorTest } = require('./error-utils')
+const routesFixture = path.join(Cypress.config('fixturesFolder'), 'routes.js')
+const appRoutesPath = path.join('app', 'routes.js')
+const appRoutes = path.join(Cypress.env('projectFolder'), appRoutesPath)
 
 const pageName = 'There is an error in your code'
 const contactSupportText = 'You can try and fix this yourself or contact the GOV.UK Prototype Kit team if you need help.'
 
 const templateNotFoundText = 'template not found: test-page.html'
 
-describe('Server Error Test', async () => {
+describe('Server Error Test', () => {
+  before(() => {
+    cy.task('log', `Replace ${appRoutes} with Cypress routes`)
+    cy.task('copyFile', { source: routesFixture, target: appRoutes })
+  })
+
+  // after(() => {
+  //   cy.task('log', `Restore ${appRoutesPath}`)
+  //   cy.task('copyFromStarterFiles', { filename: appRoutesPath })
+  // })
+
   it('internal server error results in 500 page being displayed correctly', () => {
-    setupRouterForErrorTest()
     waitForApplication()
 
     cy.visit('/error', { failOnStatusCode: false })
+
     cy.get('.govuk-heading-l').contains(pageName)
     cy.get('.govuk-body').contains(contactSupportText)
-
-    setRouterBackToInitialState()
   })
   it('shows an error if a template cannot be found', () => {
-    setupRouterForErrorTest()
     waitForApplication()
 
     cy.visit('/test-page', { failOnStatusCode: false })
+
     cy.get('.govuk-heading-l').contains(pageName)
     cy.get('.govuk-body').contains(contactSupportText)
     cy.get('code').contains(templateNotFoundText)
 
-    setRouterBackToInitialState()
+    cy.task('log', `Restore ${appRoutesPath}`)
+    cy.task('copyFromStarterFiles', { filename: appRoutesPath })
   })
 })
